+++ subs2cia/argparser.py
@@ -184,6 +184,10 @@ def get_args_subs2cia():
     parent_parser.add_argument('-a', '--absolute-paths', action='store_true', dest='absolute_paths', default=False,
                             help='Prints absolute paths from the root directory instead of given paths.')
 
+    parent_parser.add_argument('-ma', '--interactive', action='store_true', dest='interactive', default=False,
+                            help='If set, will enable interactive stream picking. Overrides -ai, -si, -tl. '
+                                 'Also overrides -c in ')
+
     cia_parser = subparsers.add_parser('condense', parents=[parent_parser],
                                        help="Condense input audio into a single audio file")
     srs_parser = subparsers.add_parser('srs', parents=[parent_parser],