+++ subs2cia/condense.py
@@ -1,4 +1,4 @@
-from subs2cia.Common import Common
+from subs2cia.Common import Common, interactive_picker
 from subs2cia.sources import AVSFile
 from subs2cia.pickers import picker
 from subs2cia.sources import Stream
@@ -31,11 +31,13 @@ from typing import Union, List
 
 
 class Condense(Common):
-    def __init__(self, sources: [AVSFile], outdir: Path, outstem: Union[str, None], condensed_video: bool, threshold: int, padding: int,
+    def __init__(self, sources: [AVSFile], outdir: Path, outstem: Union[str, None], condensed_video: bool,
+                 threshold: int, padding: int,
                  partition: int, split: int, demux_overwrite_existing: bool, overwrite_existing_generated: bool,
                  keep_temporaries: bool, target_lang: str, out_audioext: str, minimum_compression_ratio: float,
                  use_all_subs: bool, subtitle_regex_filter: str, audio_stream_index: int, subtitle_stream_index: int,
-                 ignore_range: Union[List[List[int]], None], bitrate: Union[int, None], mono_channel: bool):
+                 ignore_range: Union[List[List[int]], None], bitrate: Union[int, None], mono_channel: bool,
+                 interactive: bool):
         super(Condense, self).__init__(
             sources=sources,
             outdir=outdir,
@@ -53,7 +55,8 @@ class Condense(Common):
             subtitle_stream_index=subtitle_stream_index,
             ignore_range=ignore_range,
             bitrate=bitrate,
-            mono_channel=mono_channel
+            mono_channel=mono_channel,
+            interactive=interactive
         )
         r"""
 
@@ -72,189 +75,93 @@ class Condense(Common):
         :param minimum_compression_ratio: Chosen subtitle stream must yield generated audio at least this percent long of audio file
         """
 
-        # if outdir is None:
-        #     self.outdir = sources[0].filepath.parent
-        # else:
-        #     self.outdir = outdir
-        # self.outstem = sources[0].filepath.stem
-        # self.sources = sources
-        # self.out_audioext = out_audioext
-        # self.out_vidoeext = '.mkv'
         self.out_subext = None  # extensions must contain dot
 
-        # logging.debug(f'Will save a file with stem "{self.outstem}" to directory "{self.outdir}"')
         logging.info(f"Mapping input file(s) {sources} to one output file")
 
-        # self.partitioned_streams = defaultdict(list)
-
-        # self.picked_streams = {
-        #     'audio': None,
-        #     'subtitle': None,
-        #     'video': None
-        # }
-        #
-        # self.pickers = {
-        #     'audio': None,
-        #     'subtitle': None,
-        #     'video': None
-        # }
-
-        # self.target_lang = target_lang
-        # # following indices overrides target_lang, is overridden by standalone input audio/subtitle files
-        # self.audio_stream_index = audio_stream_index
-        # self.subtitle_stream_index = subtitle_stream_index
-
-        # self.padding = padding
         self.threshold = threshold
         self.partition = partition
         self.split = split
-        # self.ignore_range = ignore_range
 
+        self.subdata = None
         self.dialogue_times = None
         self.minimum_compression_ratio = minimum_compression_ratio
-
-        # self.demux_overwrite_existing = demux_overwrite_existing
-        # self.overwrite_existing_generated = overwrite_existing_generated
-        # self.keep_temporaries = keep_temporaries
 
         self.condensed_audio = True  # can be exposed later if needed
         self.condensed_video = condensed_video
         self.condensed_subtitles = True  # can be exposed later if needed
 
-        # self.use_all_subs = use_all_subs
-        # self.subtitle_regex_filter = subtitle_regex_filter
-
-        # self.insufficient = False
         self.subtitle_outfile = None
 
-    # # go through source files and count how many subtitle and audio streams we have
-    # def get_and_partition_streams(self):
-    #     for sourcefile in self.sources:
-    #         if sourcefile.type == 'video':
-    #             # dig into streams
-    #             for idx, stream_info in enumerate(sourcefile.info['streams']):
-    #                 stype = stream_info['codec_type']
-    #                 self.partitioned_streams[stype].append(Stream(file=sourcefile, type=stype,
-    #                                                               index=idx, stream_info=stream_info))
-    #             continue
-    #         self.partitioned_streams[sourcefile.type].append(Stream(file=sourcefile, type=sourcefile.type,
-    #                                                                 stream_info=sourcefile.info,
-    #                                                                 index=None))
-    #         # for stream in sourcefile
-    #     for k in self.partitioned_streams:
-    #         logging.info(f"Found {len(self.partitioned_streams[k])} {k} input streams")
-    #         # logging.debug(f"Streams found: {self.partitioned_streams[k]}")
+    def choose_subtitle(self, interactive: bool):
+        if len(self.partitioned_streams['subtitle']) == 0:
+            logging.warning(f"Couldn't find audio streams in input files")
+            return
+        k = 'subtitle'
+        while True:
+            if interactive and len(self.partitioned_streams['subtitle']) > 1:
+                self.picked_streams['subtitle'] = interactive_picker(self.sources, self.partitioned_streams, 'subtitle')
+            else:
 
-    # def initialize_pickers(self):
-    #     for k in self.pickers:
-    #         idx = None
-    #         if k == 'audio':
-    #             idx = self.audio_stream_index
-    #         if k == 'subtitle':
-    #             idx = self.subtitle_stream_index
-    #         self.pickers[k] = picker(self.partitioned_streams[k], target_lang=self.target_lang, forced_stream=idx)
+                k = 'subtitle'
+                try:
+                    self.picked_streams[k] = next(self.pickers[k])
+                except StopIteration as s:
+                    logging.critical(f"Input files {self.sources} don't contain usable subtitles")
+                    self.insufficient = True
+                    return
+            subfile = self.picked_streams[k].demux(overwrite_existing=self.demux_overwrite_existing)
+            if subfile is None:
+                logging.warning(f"Error while demuxing {self.picked_streams[k]}")
+                self.picked_streams[k] = None
+                continue
 
-    # def list_streams(self):
-    #     print(f"Listing streams found in {self.sources}")
-    #     for k in ['subtitle', 'audio', 'video']:
-    #         print(f"Available {k} streams:")
-    #         for idx, stream in enumerate(self.partitioned_streams[k]):
-    #             desc_str = ''
-    #             if "codec_name" in stream.stream_info:
-    #                 desc_str = desc_str + "codec: " + stream.stream_info['codec_name'] + ", "
-    #             if "tags" in stream.stream_info:
-    #                 tags = stream.stream_info['tags']
-    #                 if "language" in tags:
-    #                     desc_str = desc_str + "lang_code: " + tags['language'] + ", "
-    #                 if "title" in tags:
-    #                     desc_str = desc_str + "title: " + tags['title'] + ", "
-    #             if desc_str == '':
-    #                 desc_str = f"Stream {idx: 3}: no information found"
-    #             else:
-    #                 desc_str = f"Stream {idx: 3}: {desc_str}"
-    #             print(desc_str)
-    #         print("")
+            assert(self.picked_streams['audio'] is not None)  # must call choose_audio first
+            audiolength = subtools.get_audiofile_duration(self.picked_streams['audio'].demux_file.filepath)
+            subdata = subtools.SubtitleManipulator(subfile.filepath,
+                                                   threshold=self.threshold, padding=self.padding,
+                                                   ignore_range=self.ignore_range, audio_length=audiolength)
+            subdata.load(include_all=self.use_all_subs, regex=self.subtitle_regex_filter)
+            if subdata.ssadata is None:
+                logging.warning(f"Problem loading subtitle data from {self.picked_streams[k]}")
+                self.picked_streams[k] = None
+                continue
+            subdata.merge_groups()
+            times = subdata.get_times()
+            ps_times = subtools.partition_and_split(times, self.partition, self.split)
 
-    # def choose_streams(self):
-    #     if insufficient_source_streams(self.partitioned_streams):
-    #         logging.error(f"Not enough input sources to generate condensed output for output stem {self.outstem}")
-    #         self.insufficient = True
+            sublength = subtools.get_partitioned_and_split_times_duration(ps_times)
+
+            compression_ratio = sublength / audiolength
+            if compression_ratio < self.minimum_compression_ratio:
+                if interactive:
+                    resp = input(f"Got compression ratio of {compression_ratio} (ratio of dialogue to total audio), "
+                                 f"which is smaller than the minimum"
+                                 f" ratio of {self.minimum_compression_ratio}. Continue with picked subtitles? (y/N)")
+                    if resp != 'y':
+                        continue
+
+                else:
+                    logging.warning(f"Got compression ratio of {compression_ratio}, which is smaller than the minimum"
+                                 f" ratio of {self.minimum_compression_ratio}"
+                                 f"retrying wth different subtitle "
+                                 f"file. If too many subtitles are being ignored, try -ni or -R. If the minimum "
+                                 f"ratio is too high, try -c.")
+                    self.picked_streams[k] = None
+                    continue
+            self.subdata = subdata
+            self.dialogue_times = subtools.partition_and_split(sub_times=times,
+                                                               partition_size=1000 * self.partition,
+                                                               split_size=1000 * self.split)
+            break
+
+
+    # def process_subtitles(self):
+    #     if self.picked_streams['subtitle'] is None:
+    #         logging.error(f'No subtitle stream to process for output stem {self.outstem}')
     #         return
-    #     while picked_sources_are_insufficient(self.picked_streams):
-    #         for k in ['audio', 'video', 'subtitle']:
-    #             if len(self.partitioned_streams[k]) == 0:
-    #                 logging.debug(f"no input streams of type {k}")
-    #                 continue
-    #             if self.picked_streams[k] is None:
-    #                 try:
-    #                     self.picked_streams[k] = next(self.pickers[k])
-    #                 except StopIteration as s:
-    #                     logging.critical("Input streams for this group are invalid for condensing")
-    #                     self.insufficient = True
-    #                     return
-    #         for k in ['audio', 'subtitle', 'video']:
-    #             # validate picked streams
-    #
-    #             # todo: spin off into its own function at a later step
-    #             if k == 'audio':
-    #                 afile = self.picked_streams[k].demux(overwrite_existing=self.demux_overwrite_existing)
-    #                 if afile is None:
-    #                     self.picked_streams[k] = None
-    #
-    #             if k == 'subtitle':
-    #                 subfile = self.picked_streams[k].demux(
-    #                     overwrite_existing=self.demux_overwrite_existing)  # type AVSFile
-    #                 if subfile is None:
-    #                     self.picked_streams[k] = None
-    #                     continue
-    #                 # times = subtools.load_subtitle_times(subfile.filepath, include_all_lines=self.use_all_subs)
-    #                 subdata = subtools.SubtitleManipulator(subfile.filepath,
-    #                                                        threshold=self.threshold, padding=self.padding,
-    #                                                        ignore_range=self.ignore_range)
-    #                 subdata.load(include_all=self.use_all_subs, regex=self.subtitle_regex_filter)
-    #                 if subdata.ssadata is None:
-    #                     self.picked_streams[k] = None
-    #                     continue
-    #                 if self.picked_streams['audio'] is None:
-    #                     # can't verify subtitle validity until audio candidate is found
-    #                     continue
-    #                 # times = subtools.merge_times(times, threshold=self.threshold, padding=self.padding)
-    #                 subdata.merge_groups()
-    #                 times = subdata.get_times()
-    #                 ps_times = subtools.partition_and_split(times, self.partition, self.split)
-    #
-    #                 sublength = subtools.get_partitioned_and_split_times_duration(ps_times)
-    #                 audiolength = subtools.get_audiofile_duration(
-    #                     self.picked_streams['audio'].demux_file.filepath)
-    #                 compression_ratio = sublength / audiolength
-    #                 if compression_ratio < self.minimum_compression_ratio:
-    #                     logging.info(f"got compression ratio of {compression_ratio}, which is smaller than the minimum"
-    #                                  f" ratio of {self.minimum_compression_ratio}, retrying wth different subtitle "
-    #                                  f"file. If too many subtitles are being ignored, try -ni or -R. If the minimum "
-    #                                  f"ratio is too high, try -c.")
-    #                     self.picked_streams[k] = None
-    #                     continue
-    #                 self.dialogue_times = subtools.partition_and_split(sub_times=times,
-    #                                                                    partition_size=1000 * self.partition,
-    #                                                                    split_size=1000 * self.split)
-    #             if k == 'video':
-    #                 pass
-    #     logging.info(f"Picked {self.picked_streams['audio']} to use for condensing")
-    #     logging.info(f"Picked {self.picked_streams['video']} to use for condensing")
-    #     logging.info(f"Picked {self.picked_streams['subtitle']} to use for condensing")
-
-    def process_subtitles(self):
-        if self.picked_streams['subtitle'] is None:
-            logging.error(f'No subtitle stream to process for output stem {self.outstem}')
-            return
-        if self.insufficient:
-            return
-        logging.debug("process_subtitles merged into choose_streams")
-        # subfile = self.picked_streams['subtitle'].demux(overwrite_existing=self.demux_overwrite_existing)
-        # times = subtools.load_subtitle_times(subfile.filepath)
-        # times = subtools.merge_times(times, threshold=self.threshold, padding=self.padding)
-        # self.dialogue_times = subtools.partition_and_split(sub_times=times, partition_size=1000*self.partition,
-        #                                                    split_size=1000*self.split)
+    #     if self.insufficient:
+    #         return
 
     def export_subtitles(self):
         if self.picked_streams['subtitle'] is None:
@@ -262,14 +169,11 @@ class Condense(Common):
             return
         subpath = self.picked_streams['subtitle'].get_data_path()
         subext = subpath.suffix
-        subdata = subtools.SubtitleManipulator(subpath, threshold=self.threshold, padding=self.padding, ignore_range=self.ignore_range, audio_length=subtools.get_audiofile_duration(self.picked_streams['audio'].demux_file.filepath))
-        subdata.load(include_all=self.use_all_subs, regex=self.subtitle_regex_filter)
-        subdata.merge_groups()
-        subdata.condense()
+        self.subdata.condense()
 
         self.subtitle_outfile = Path(self.outdir) / (
                     self.outstem + f'.condensed{self.out_subext if self.out_subext is not None else subext}')
-        subdata.condensed_ssadata.save(self.subtitle_outfile, encoding=u'utf-8')
+        self.subdata.condensed_ssadata.save(self.subtitle_outfile, encoding=u'utf-8')
 
     def export_audio(self):
         if self.picked_streams['audio'] is None: